+++ scripts/Install_PycharmCommunity.ps1
@@ -3,14 +3,22 @@ $Install ={
     {
         param (
             [string] $url,
-            [string] $dest_folder_exe
+            [string] $prog_files_dest,
+            [string] $source_exe_location
         )
 
         Write-Output 'Installing...'
 
-        # Start downloading
-        Import-Module BitsTransfer
-        Start-BitsTransfer -Source $url -Destination $dest_folder_exe
+        if (!(Test-Path $source_exe_location))
+        {
+            # Start downloading
+            Import-Module BitsTransfer
+            Start-BitsTransfer -Source $url -Destination $prog_files_dest
+        }
+        else
+        {
+            Write-Output 'Pycharm exe file already exists on your computer'
+        }
     }
 }
 
@@ -29,14 +37,24 @@ $Shortcut ={
         $shortcut_dest_location = $DesktopPath + '\pycharm-community.lnk'
         $source_icon_location = $dest_directory + '\bin\pycharm64.exe'
 
-        # Create shortcut on desktop
-        $WScriptShell = New-Object -ComObject WScript.Shell
-        $Shortcut = $WScriptShell.CreateShortcut($shortcut_dest_location)
-        $Shortcut.TargetPath = $source_icon_location
-        $Shortcut.Save()
+        if (!(Test-Path $shortcut_dest_location))
+        {
+            # Create shortcut on desktop
+            $WScriptShell = New-Object -ComObject WScript.Shell
+            $Shortcut = $WScriptShell.CreateShortcut($shortcut_dest_location)
+            $Shortcut.TargetPath = $source_icon_location
+            $Shortcut.Save()
+        }
+        else
+        {
+            Write-Output 'Shortcut icon already exists on desktop'
+        }
 
-        # Remove exe file from computer
-        Remove-Item $source_exe_location
+        if (Test-Path $source_exe_location)
+        {
+            # Remove exe file from computer
+            Remove-Item $source_exe_location
+        }
 
         Write-Output 'Your folder directory located:'$dest_directory
         Write-Output 'Done installing pycharm on your computer and create shortcut on desktop'
@@ -50,24 +68,30 @@ $url = 'https://download-cf.jetbrains.com/python/pycharm-community-2020.1.exe'
 $file_name = [System.IO.Path]::GetFileName($url)
 
 # Destination folder on computer
-$dest_folder_exe = [Environment]::GetEnvironmentVariable('ProgramFiles');
+$prog_files_dest = [Environment]::GetEnvironmentVariable('ProgramFiles');
 
 # Source of exe file location
-$source_exe_location= $Env:Programfiles + '\' + $file_name
+$source_exe_location= $prog_files_dest+ '\' + $file_name
 
 # Create a new folder name for exe data destination
 $folder_data_name = [string]$file_name.TrimEnd('.exe')
 
 # Installation destination
-$dest_directory = $Env:Programfiles+'\'+$folder_data_name
+$dest_directory = $prog_files_dest+'\'+$folder_data_name
 
 # Config file for install
 $location_config= $PSScriptRoot
 $config_file = $location_config +'\silent_pycharm.config'
 
-$startProc = Start-Process powershell -Verb runAs -PassThru -ArgumentList "-NoProfile -ExecutionPolicy Bypass -Command & {$Install run '$url' '$dest_folder_exe'}"
-$startProc.WaitForExit()
+Start-Process -Wait powershell -Verb runAs -PassThru -ArgumentList "-NoProfile -ExecutionPolicy Bypass -Command & {$Install run '$url' '$prog_files_dest' '$source_exe_location'}"
 
-Start-Process -Wait -FilePath $source_exe_location -Argument "/S /CONFIG=$config_file /D=$dest_directory"
+if (!(Test-Path $dest_directory))
+{
+    Start-Process -Wait -FilePath $source_exe_location -Argument "/S /CONFIG=$config_file /D=$dest_directory" -PassThru
+}
+else
+{
+    Write-Output 'Pycharm directory already exists on your computer'
+}
 
 Start-Process -Wait powershell -Verb runAs -PassThru -ArgumentList "-NoProfile -NoExit -ExecutionPolicy Bypass -Command & {$Shortcut run '$source_exe_location' '$dest_directory'}"
